+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add hook method `AbstractReplicantEndpoint.onSessionOpen(...)`
+
 ### [v6.134](https://github.com/replicant4j/replicant/tree/v6.134) (2024-07-19) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.133...v6.134)
 
 Changes in this release: